rename&restructure: GNB 폴더구조 변경
