+++ pages/_app.tsx
@@ -2,7 +2,7 @@ import type { AppProps } from 'next/app';
 import { SpeedInsights } from '@vercel/speed-insights/next';
 import { Analytics } from '@vercel/analytics/react';
 import Head from 'next/head';
-import GNB from '@/widgets/gnb';
+import GNB from '@/widgets/GNB';
 import Provider from '@/app/providers';
 import './styles/globals.css';
