+++ src/widgets/gnb/index.tsx
@@ -1,7 +1,7 @@
 import { PropsWithChildren } from 'react';
 import { Noto_Sans_KR } from 'next/font/google';
-import GNBHeader from '@/widgets/gnb/header';
-import GNBFooter from '@/widgets/gnb/footer';
+import GNBHeader from './ui/GNB_header';
+import GNBFooter from './ui/GNB_footer';
 import useAuth from '@/shared/hooks/useAuth';
 
 const noto = Noto_Sans_KR({
