+++ src/widgets/gnb/ui/GNB_footer/index.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import { hideGnbFooterRoutes, menuItems } from '@/shared/constants/gnb';
-import GNBItem from '../item';
+import GNBItem from '../GNB_item';
 import { useRouter } from 'next/router';
 import { useSelector } from 'react-redux';
 import { RootState } from '@/shared/lib/redux/store';
