+++ src/widgets/gnb/ui/GNB_item/wrapper/index.tsx
